Remove the dependency on Quaternions.jl (#38)
Remove the dependency on [Quaternions.jl](https://github.com/JuliaGeometry/Quaternions.jl), and get rid of `REQUIRE` too.

The functions involving rotations and quaternions are now implemented internally. They should be a bit faster than the ones provides in `Quaternions.jl`, as they assume that the quaternions are always normalized.